Move fledge logic to fledge module
