+++ modules/fledgeForGpt.js
@@ -4,7 +4,7 @@
  */
 import { config } from '../src/config.js';
 import { getHook } from '../src/hook.js';
-import { getGptSlotForAdUnitCode, logInfo, logWarn } from '../src/utils.js';
+import { getGptSlotForAdUnitCode, logInfo, logWarn, memoize } from '../src/utils.js';
 
 const MODULE = 'fledgeForGpt'
 
@@ -50,3 +50,14 @@ export function addComponentAuctionHook(next, bidRequest, componentAuctionConfig
 
   next(bidRequest, componentAuctionConfig);
 }
+
+export function markBidsForFledge(next, bidderRequests) {
+  if (navigator.runAdAuction) {
+    bidderRequests.forEach((req) => {
+      req.fledgeEnabled = config.runWithBidder(req.bidderCode, () => config.getConfig('fledgeEnabled'))
+    })
+  }
+  next(bidderRequests);
+}
+
+getHook('makeBidRequests').after(markBidsForFledge);
