+++ src/adapterManager.js
@@ -321,13 +321,6 @@ adapterManager.makeBidRequests = hook('sync', function (adUnits, auctionStart, a
     });
   }
 
-  bidRequests.forEach(bidRequest => {
-    config.runWithBidder(bidRequest.bidderCode, () => {
-      const fledgeEnabledFromConfig = config.getConfig('fledgeEnabled');
-      bidRequest['fledgeEnabled'] = navigator.runAdAuction && fledgeEnabledFromConfig
-    });
-  });
-
   return bidRequests;
 }, 'makeBidRequests');
